requirements.txt
python-dotenv==0.12.0
praw==7.1.0